+++ bauh/view/core/config.py
@@ -41,7 +41,8 @@ def read_config(update_file: bool = False) -> dict:
             'hdpi': True,
             "auto_scale": False,
             "scale_factor": 1.0,
-            'stylesheet': 'default'
+            'theme': 'default-light',
+            'system_theme': True
 
         },
         'download': {